+++ MicrofluidSimulator/MicrofluidSimulator/SimulatorCode/Models/HeaterActuatorModels.cs
@@ -9,7 +9,7 @@ namespace MicrofluidSimulator.SimulatorCode.Models
         public static ArrayList heaterTemperatureChange(Container container, Heater heater)
         {
             heater.SetPowerStatus();
-            heater.valueActualTemperature = (heater.valueActualTemperature + container.timeStep * heater.valuePowerStatus);
+            heater.valueActualTemperature = (heater.valueActualTemperature + (float) container.timeStep * heater.valuePowerStatus);
             
             return heater.subscriptions;
